added removeMarker function to allow single marker removal
Currently, the only way to remove a marker in gmap is to remove all the
markers.  This can be a bit of a drawback when you've got something like
an interactive map where users can add and remove markers at will.  So I
added a removeMarker which you can pass the google maps Marker object
and it will remove the marker and markerKey (if it exists).